+++ tests/publisher/conftest.py
@@ -5,10 +5,7 @@ from pytest import MonkeyPatch
 
 from mex.common.backend_api.connector import BackendApiConnector, ReferenceFilter
 from mex.common.models import (
-    MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
-    AnyExtractedModel,
     AnyMergedModel,
-    ExtractedPrimarySource,
     ItemsContainer,
     MergedBibliographicResource,
     MergedConsent,
@@ -16,12 +13,13 @@ from mex.common.models import (
     MergedOrganizationalUnit,
     MergedPerson,
     MergedPrimarySource,
-    PaginatedItemsContainer,
+    VersionStatus,
 )
 from mex.common.types import (
     AccessRestriction,
     MergedOrganizationalUnitIdentifier,
     MergedPersonIdentifier,
+    Text,
     YearMonthDayTime,
 )
 
@@ -118,6 +116,35 @@ def merged_unit_contactpoint_container() -> ItemsContainer[AnyMergedModel]:
 
 
 @pytest.fixture
+def merged_bibliographic_resource_list() -> list[MergedBibliographicResource]:
+    """Mock a list of Merged Bibliographic Resource items."""
+    return [
+        MergedBibliographicResource(
+            accessRestriction=AccessRestriction["OPEN"],
+            creator=["PersonIdentifier"],
+            identifier="PublicationOfC1",
+            title=[Text(value="title 1, Unit C1", language=None)],
+            contributingUnit=["6rqNvZSApUHlz8GkkVP48"],  # C1
+        ),
+        MergedBibliographicResource(
+            accessRestriction=AccessRestriction["OPEN"],
+            creator=["PersonIdentifier"],
+            identifier="PublicationOfPRNTUnit",
+            publicationYear="2042",
+            title=[Text(value="title 1, Unit Parent", language=None)],
+            contributingUnit=["hIiJpZXVppHvoyeP0QtAoS"],  # PRNT
+        ),
+        MergedBibliographicResource(
+            accessRestriction=AccessRestriction["OPEN"],
+            creator=["PersonIdentifier"],
+            identifier="PublicationOfFG99",
+            title=[Text(value="title 1, Unit FG99", language=None)],
+            contributingUnit=["cjna2jitPngp6yIV63cdi9"],  # FG99
+        ),
+    ]
+
+
+@pytest.fixture
 def mocked_publisher_fallback_unit_identifiers_by_person() -> dict[
     MergedPersonIdentifier, list[MergedOrganizationalUnitIdentifier]
 ]:
@@ -126,29 +153,6 @@ def mocked_publisher_fallback_unit_identifiers_by_person() -> dict[
             MergedOrganizationalUnitIdentifier("ValidUnitWithEmail")
         ]
     }
-
-
-def fetch_merged_items(  # noqa: PLR0913
-    *,
-    query_string: str | None = None,  # noqa: ARG001
-    identifier: str | None = None,  # noqa: ARG001
-    entity_type: list[str] | None = None,
-    reference_filters: list[ReferenceFilter] | None = None,  # noqa: ARG001
-    skip: int = 0,  # noqa: ARG001
-    limit: int = 100,  # noqa: ARG001
-) -> PaginatedItemsContainer[AnyMergedModel]:
-    merged_items: list[AnyMergedModel] = [
-        MergedContactPoint(
-            email=["mex@rki.de"],
-            identifier="fakeFakeContact",
-        ),
-    ]
-    items = [
-        item
-        for item in merged_items
-        if not entity_type or item.entityType in entity_type
-    ]
-    return PaginatedItemsContainer[AnyMergedModel](total=len(items), items=items)
 
 
 def fetch_all_publishable_merged_items(
@@ -197,45 +201,12 @@ def fetch_all_publishable_merged_items(
     ]
 
 
-def fetch_extracted_items(
-    *,
-    query_string: str | None = None,  # noqa: ARG001
-    stable_target_id: str | None = None,  # noqa: ARG001
-    entity_type: list[str] | None = None,
-    skip: int = 0,  # noqa: ARG001
-    limit: int = 100,  # noqa: ARG001
-) -> PaginatedItemsContainer[AnyExtractedModel]:
-    extracted_items: list[AnyExtractedModel] = [
-        ExtractedPrimarySource(
-            hadPrimarySource=MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
-            identifierInPrimarySource="wikidata",
-        ),
-        ExtractedPrimarySource(
-            hadPrimarySource=MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
-            identifierInPrimarySource="endnote",
-        ),
-    ]
-    items = [
-        item
-        for item in extracted_items
-        if not entity_type or item.entityType in entity_type
-    ]
-    return PaginatedItemsContainer[AnyExtractedModel](total=len(items), items=items)
-
-
 @pytest.fixture
-def mocked_backend(monkeypatch: MonkeyPatch) -> MagicMock:
+def mocked_backend_publisher(monkeypatch: MonkeyPatch) -> MagicMock:
     backend = MagicMock(
-        fetch_merged_items=MagicMock(
-            spec=BackendApiConnector.fetch_merged_items, side_effect=fetch_merged_items
-        ),
         fetch_all_publishable_merged_items=MagicMock(
             spec=BackendApiConnector.fetch_all_publishable_merged_items,
             side_effect=fetch_all_publishable_merged_items,
-        ),
-        fetch_extracted_items=MagicMock(
-            spec=BackendApiConnector.fetch_extracted_items,
-            side_effect=fetch_extracted_items,
         ),
     )
     monkeypatch.setattr(
@@ -243,20 +214,16 @@ def mocked_backend(monkeypatch: MonkeyPatch) -> MagicMock:
     )
     monkeypatch.setattr(
         BackendApiConnector,
-        "request",
+        "system_status",
         MagicMock(
-            return_value={"status": "Fabulous", "version": "mex-backend-version"}
+            return_value=VersionStatus.model_validate(
+                {"status": "Fabulous", "version": "mex-backend-version"}
+            )
         ),
-    )
-    monkeypatch.setattr(
-        BackendApiConnector, "fetch_merged_items", backend.fetch_merged_items
     )
     monkeypatch.setattr(
         BackendApiConnector,
         "fetch_all_publishable_merged_items",
         backend.fetch_all_publishable_merged_items,
-    )
-    monkeypatch.setattr(
-        BackendApiConnector, "fetch_extracted_items", backend.fetch_extracted_items
     )
     return backend